fix(workflows): sidebar was not displayed sometimes when refreshing
Based in my research, is because we are using a Portal to mount the
sidebar items.
The problem comes when sometimes the sidebar.tsx was mounted after
the portal, which is a problem because the Portal cannot find that
element in the DOM, and then doesn't render the options.

You can see how that is defined in the Portal component.

Trigger option checks continuously is the DOM element where the portal
will be mounted exists, and that is exactly what I want. If the
Portal is mounted first, do a polling to see if the element was mounted
after.

I were using that logic for mobile menu, so that is the reason why
I'm keeping that option if we detect that the user is in a mobile, and
just do the trigger by default in desktop.